requirements.txt
praw==7.7.1
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.4
openai==1.12.0
nltk==3.8.1
textblob==0.17.1 